update current branch with its remote counterpart
useful for when you are amending a current branch

Changes:

GitPath
* updated get_all_branches to return list

GetCurrentCommitInfo
* Changed from get_all_files_from_commit to get_current_commit_info
  instead.
* Old class was GetAllFilesFromCommit
* Since get_current_changes_as_arrays() is going to be used in different
  spots, I placed it within this class. Note that this uses
  get_delta_files() function as well!

GitBranchShortcut
* Refactored checkout and delete shortcut functions
* Created git_update_current_branch() function
* Allowing user to update their current branch with its remote
  counterpart in case it falls out of sync.

gs.py
* Took off warning and version print from all iterations and only
  printing them in startup now.
* Replaced mikkel_merge call with update branch
