+++ compile.bat
@@ -1,0 +1,1 @@
+gcc main.c -o gsshortcut.exe
