+++ rebase_branches.json
@@ -1,1 +1,1 @@
-{"rebase_branches": {"origin/main": true}}
+{"rebase_branches": {"origin/main": true, "origin/master": true}}
